requirements.txt
python-telegram-bot>=21.0
requests==2.31.0
beautifulsoup4==4.12.2
pyyaml==6.0.1 